29-4 previo a publicar

+++ src/app/components/educacion/listareducacion/listareducacion.component.ts
@@ -1,8 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { Educacion } from 'src/app/Modelo/Educacion';
 import { EducacionService } from 'src/app/Service/educacion.service';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
+import { TokenService } from 'src/app/Service/token.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-listareducacion',
@@ -13,22 +15,25 @@ export class ListareducacionComponent implements OnInit {
 
   listaEducacion: Educacion[];
   usuario: string;
-  visible: boolean;
+  visible: boolean = false;
+  
 
-  constructor(private serviceEduc: EducacionService, private router: Router) { }
+  constructor(private serviceEduc: EducacionService, private router: Router, private tokenService: TokenService) { }
 
   ngOnInit(): void {
-     this.serviceEduc.getlistaEducacion()
+    this.serviceEduc.getlistaEducacion()
       .subscribe(data => { this.listaEducacion = data; })
-      
 
-      //login
-      this.usuario = localStorage.getItem("user");
-      if (this.usuario == "admin") {
-        this.visible = true;
-      } else {
-        this.visible = false;
-      }
+    this.visible = this.tokenService.isAdmin();
+
+
+    //login previo
+    /* this.usuario = localStorage.getItem("user");
+    if (this.usuario == "admin") {
+      this.visible = true;
+    } else {
+      this.visible = false;
+    } */
 
   }
 
@@ -38,18 +43,35 @@ export class ListareducacionComponent implements OnInit {
     this.router.navigate(["add-educacion"]);
   }
 
-//NO lo uso con routerlink
+  //NO lo uso con routerlink
   Editar(objEduc: Educacion): void {
     localStorage.setItem("id_educacion", objEduc.id_educacion.toString());
     this.router.navigate(["edit-educacion"]);
   }
 
   Delete(objEduc: Educacion) {
-    this.serviceEduc.deleteEducacion(objEduc)
-       .subscribe(data => {
-        this.listaEducacion = this.listaEducacion.filter(p => p !== objEduc);
-        alert("Estudio eliminando....");
-      })
+    Swal.fire({
+      title: 'Esta seguro?',
+      text: "Esta a punto de eliminar un registro!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        this.serviceEduc.deleteEducacion(objEduc)
+          .subscribe(data => {
+            this.listaEducacion = this.listaEducacion.filter(p => p !== objEduc);
+            Swal.fire(
+              'Eliminado!',
+              'El registro ha sido eliminado.',
+              'success'
+            )
+          })
+      }
+    })
+    
   }
 
   onDropped(event: CdkDragDrop<any>) {
