+++ src/app/components/experiencia/listarexperiencia/listarexperiencia.component.ts
@@ -1,8 +1,11 @@
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { ExperienciaLaboral } from 'src/app/Modelo/ExperienciaLaboral';
 import { ExperiencialaboralService } from 'src/app/Service/experiencialaboral.service';
+import { TokenService } from 'src/app/Service/token.service';
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: 'app-listarexperiencia',
@@ -13,23 +16,18 @@ export class ListarexperienciaComponent implements OnInit {
 
   listaExpLaboral: ExperienciaLaboral[];
   usuario: string;
-  visible: boolean;
+  visible: boolean = false;
 
-  constructor(private serviceExpLaboral: ExperiencialaboralService, private router: Router) { }
+  constructor(private serviceExpLaboral: ExperiencialaboralService, private router: Router, private tokenService: TokenService) { }
 
   ngOnInit(): void {
-     this.serviceExpLaboral.getlistaExpLaboral()
+    this.serviceExpLaboral.getlistaExpLaboral()
       .subscribe(data => { this.listaExpLaboral = data; })
 
-      this.usuario = localStorage.getItem("user");
-      if (this.usuario == "admin") {
-        this.visible = true;
-      } else {
-        this.visible = false;
-      }
+    this.visible = this.tokenService.isAdmin();
 
   }
- 
+
   Nuevo() {
     this.router.navigate(["add-explaboral"]);
   }
@@ -39,13 +37,31 @@ export class ListarexperienciaComponent implements OnInit {
     this.router.navigate(["edit-explaboral"]);
   }
 
+
   Delete(objExplab: ExperienciaLaboral) {
-    this.serviceExpLaboral.deleteExpLaboral(objExplab)
-       .subscribe(data => {
-        this.listaExpLaboral = this.listaExpLaboral.filter(p => p !== objExplab);
-        alert("Estudio eliminando....");
-      })
+    Swal.fire({
+      title: 'Esta seguro?',
+      text: "Esta a punto de eliminar un registro!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        this.serviceExpLaboral.deleteExpLaboral(objExplab)
+          .subscribe(data => {
+            this.listaExpLaboral = this.listaExpLaboral.filter(p => p !== objExplab);
+            Swal.fire(
+              'Eliminado!',
+              'El registro ha sido eliminado.',
+              'success'
+            )
+          })
+      }
+    })
   }
+
 
   onDropped(event: CdkDragDrop<any>) {
     moveItemInArray(this.listaExpLaboral, event.previousIndex, event.currentIndex);
