+++ src/app/components/proyectos/listarproyecto/listarproyecto.component.ts
@@ -3,6 +3,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Proyectos } from 'src/app/Modelo/Proyectos';
 import { ProyectoService } from 'src/app/Service/proyecto.service';
+import { TokenService } from 'src/app/Service/token.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-listarproyecto',
@@ -13,25 +15,20 @@ export class ListarproyectoComponent implements OnInit {
 
   listaProyectos: Proyectos[];
   usuario: string;
-  visible: boolean;
+  visible: boolean = false;
 
-  constructor(private serviceProy: ProyectoService, private router: Router) { }
+
+  constructor(private serviceProy: ProyectoService, private router: Router, private tokenService: TokenService) { }
 
   ngOnInit(): void {
     this.serviceProy.getlistaProyectos()
       .subscribe(data => { this.listaProyectos = data; })
 
-    this.usuario = localStorage.getItem("user");
-    if (this.usuario == "admin") {
-      this.visible = true;
-    } else {
-      this.visible = false;
-    }
+    this.visible = this.tokenService.isAdmin();
 
   }
 
 
-  //metodo para acceder a la ruta del componente AddEducacionComponent
   Nuevo() {
     this.router.navigate(["add-proyecto"]);
   }
@@ -43,13 +40,30 @@ export class ListarproyectoComponent implements OnInit {
   }
 
   Delete(objProy: Proyectos) {
-    this.serviceProy.deleteProyecto(objProy)
-      .subscribe(data => {
-        this.listaProyectos = this.listaProyectos.filter(p => p !== objProy);
-        alert("Proyecto eliminado....");
-      })
+    Swal.fire({
+      title: 'Esta seguro?',
+      text: "Esta a punto de eliminar un registro!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        this.serviceProy.deleteProyecto(objProy)
+          .subscribe(data => {
+            this.listaProyectos = this.listaProyectos.filter(p => p !== objProy);
+            Swal.fire(
+              'Eliminado!',
+              'El registro ha sido eliminado.',
+              'success'
+            )
+          })
+      }
+    })
   }
 
+  
   onDropped(event: CdkDragDrop<any>) {
     moveItemInArray(this.listaProyectos, event.previousIndex, event.currentIndex);
   }
